Removing MapFileReadResult by folding it into MapReadResult #668
