--- mapsforge-map-reader/src/main/java/org/mapsforge/map/reader/PoiWayBundle.java
@@ -1,30 +1,0 @@
-/*
- * Copyright 2010, 2011, 2012, 2013 mapsforge.org
- *
- * This program is free software: you can redistribute it and/or modify it under the
- * terms of the GNU Lesser General Public License as published by the Free Software
- * Foundation, either version 3 of the License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful, but WITHOUT ANY
- * WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A
- * PARTICULAR PURPOSE. See the GNU Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public License along with
- * this program. If not, see <http://www.gnu.org/licenses/>.
- */
-package org.mapsforge.map.reader;
-
-import org.mapsforge.map.datastore.PointOfInterest;
-import org.mapsforge.map.datastore.Way;
-
-import java.util.List;
-
-public class PoiWayBundle {
-	final List<PointOfInterest> pois;
-	final List<Way> ways;
-
-	public PoiWayBundle(List<PointOfInterest> pois, List<Way> ways) {
-		this.pois = pois;
-		this.ways = ways;
-	}
-}
