+++ mapsforge-map-reader/src/test/java/org/mapsforge/map/reader/EncodingTest.java
@@ -18,6 +18,7 @@ import org.junit.Assert;
 import org.mapsforge.core.model.LatLong;
 import org.mapsforge.core.model.Tile;
 import org.mapsforge.core.util.MercatorProjection;
+import org.mapsforge.map.datastore.MapReadResult;
 import org.mapsforge.map.datastore.Way;
 
 final class EncodingTest {
@@ -29,7 +30,7 @@ final class EncodingTest {
 		int tileY = MercatorProjection.latitudeToTileY(0, ZOOM_LEVEL);
 		Tile tile = new Tile(tileX, tileY, ZOOM_LEVEL, 256);
 
-		MapFileReadResult mapReadResult = mapFile.readMapData(tile);
+		MapReadResult mapReadResult = mapFile.readMapData(tile);
 		mapFile.close();
 
 		Assert.assertTrue(mapReadResult.pointOfInterests.isEmpty());
